added config file that saves user options, added options for selecting folder path to osu/osu!lazer + added option to select which client user want replay from

+++ OsuFileParser/Classes/Beatmap/osu/BeatmapClasses/TimingPoints.cs
@@ -2,7 +2,7 @@
 {
     public class TimingPoint
     {
-        public int Time { get; set; }
+        public decimal Time { get; set; }
         public decimal BeatLength { get; set; }
         public int Meter { get; set; }
         public int SampleSet { get; set; }
